changed usigned long declaration to uint32_t

+++ include/n200-driver/n200_timer.h
@@ -4,7 +4,7 @@
 #define N200_TIMER_H
 
 /** SystemCoreClock has to be set by the user */
-extern volatile unsigned long SystemCoreClock;
+extern volatile uint32_t SystemCoreClock;
 
 #define TIMER_MSIP 0xFFC
 #define TIMER_MSIP_size   0x4
